refactor(org-admin): split attendees page under the line cap (#544)
Extract the attendees admin page (1016 lines) into cohesive sub-components
under $lib/components/attendees/, mirroring the tickets page structure:
AttendeeStats, AttendeeFilters, AttendeeTable (desktop), AttendeeCardList
(mobile), and EditRsvpDialog. Queries, mutations, and URL/filter state stay
in the page; only rendering moves behind narrow props. Zero behavior change.
